Rename files to match component name
