+++ src/DetailsView/reports/components/report-sections/rules-only.tsx
